Remove dependence on inheritance
